LayoutTests/fast/loader/stateobjects/pushstate-with-fragment-urls-and-hashchange.html

<html>
<head>
<script>

if (window.layoutTestController) {
    layoutTestController.clearBackForwardList();
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
}

function log(txt)
{
    document.getElementById("logger").innerText += txt + "\n";
}

function runTest()
{
    history.replaceState("OriginalEntry", "original");
    history.pushState(null, null, "#");
    history.pushState(null, null, "");
    history.pushState(null, null, "#hash");
    history.pushState(null, null, "#otherhash");
    history.pushState(null, null, "#");
    history.pushState(null, null, null);
    history.pushState(null, null, "some-other.html");
    history.pushState(null, null, "some-other.html?withquery");
    history.pushState(null, null, "some-other.html?withquery#");
    history.pushState(null, null, "some-other.html?withquery");
    history.pushState(null, null, "some-other.html?withquery#somehash");
    history.pushState(null, null, "some-other.html?withquery#");
    history.pushState(null, null, "some-other.html?withquery#someotherhash");
    history.pushState(null, null, "some-other.html?withsomeotherquery#someotherhash");
    history.pushState(null, null, "some-other.html?withsomeotherquery#somehash");
    history.pushState(null, null, "some-other.html?withsomeotherquery");
    history.pushState(null, null, "some-other.html?withsomeotherquery#");
    history.pushState(null, null, "some-other.html?withsomeotherquery");

    history.pushState("BufferEntry", "Last entry");
    history.back();
}

function statePopped()
{
    log("State popped with event " + event.state + " (type " + typeof event.state + ") and location " + window.location);
    if (event.state != "OriginalEntry")
        history.back();
    else if (window.layoutTestController)
        layoutTestController.notifyDone();
}

function hashChanged()
{
    log("Hash change fired");
}

</script>
<body onload="runTest();" onpopstate="statePopped();" onhashchange="hashChanged();">
<pre>
This test pushes a series of state objects with different URLs and fragment identifiers meant to test the hashChange event as states are popped.
</pre><br>
<pre id="logger"></pre>
</body>
</html>
